Move preloaded function implementations
Simple redirection functions are moved into the libc preload lib,
more substantial implementations (getaddrinfo and getifaddrs) are
moved into shim_api_* modules that are called from the libc
preload lib.
